+++ board/rbz/common/eeprom_hw.h
@@ -2,6 +2,6 @@
 #define __EEPROM_HW_H_
 
 int board_read_rom_eeprom(unsigned char *config);
-u32 get_ramsize(unsigned char config);
+long get_ramsize(unsigned char config);
 
 #endif
